+++ src/Nethermind/Nethermind.State/IVisitingWorldState.cs
@@ -9,12 +9,4 @@ namespace Nethermind.State;
 
 public interface IVisitingWorldState : IWorldState
 {
-    /// <summary>
-    /// Runs a visitor over trie.
-    /// </summary>
-    /// <param name="visitor">Visitor to run.</param>
-    /// <param name="stateRoot">Root to run on.</param>
-    /// <param name="visitingOptions">Options to run visitor.</param>
-    void Accept<TCtx>(ITreeVisitor<TCtx> visitor, Hash256 stateRoot, VisitingOptions? visitingOptions = null)
-        where TCtx : struct, INodeContext<TCtx>;
 }
